backend/requirements.txt
fastapi
uvicorn[standard]
crawl4ai
httpx
python-dotenv
